Create HRY and Abbas
链接：https://ac.nowcoder.com/acm/contest/874/C
来源：牛客网

In a parallel universe, HRY is the president of the Kassel Academy Student Union. Once he and the Lionheart Association President Abdullah Abbas performed a secret mission: to investigate the mixed blood drug trafficking group in country X.

However, an accident happened and they were caught together. They were forced to play Russian gambling roulette games!

The terrorists brought a revolver which works as follows: a revolver has n places 1,2,...,n to place bullet, and these n places forms a circle. The revolver has a position it is currently pointing at, and when pulling the trigger, the revolver will shoot out the bullet (if it has) at the current position and turn to the next position. The next position of 
1
≤
i
≤
n
−
1
1≤i≤n−1 is i+1, and for n the next position is 1. Now the terrorist puts m bullets in the revolver, and turned the wheel of the revolver, which means the starting position is randomly chosen. He orders HRY and Abbas shoot at their heads in turn using the same revolver. If someone is hit by bullet, then the game is over.

HRY and Abbas are both not ordinary mixed blood, so the revolver can do no harm to them. However, they still decided to play the game. Now they want to know the probability of the game ending exactly after the ith shoot.

输入描述:
The first line is an integer T, indicating the number of test cases. 
For each test case there are two lines :
The first line contains two integers n,m, and the second line contains m different positive integers 
a
i
ai, indicating the initial positions of bullets.
1
≤
T
≤
1000
,
1
≤
m
≤
n
≤
10
5
,
1
≤
a
i
≤
n
.
1≤T≤1000,1≤m≤n≤105,1≤ai≤n.
The sum of n does not exceed 
10
6
106.
输出描述:
For each test case output n lines, the ith line indicates the probability of game ending exactly after the ith shoot. The probability is in the form of irreducible fraction p/q, which means the greatest common divisor of p and q is 1. If the answer is 0, output 0 directly.
示例1
输入
复制
2
10 2
1 3
10 4
1 2 6 10
输出
复制
1/5
1/5
1/10
1/10
1/10
1/10
1/10
1/10
0
0
2/5
1/5
1/5
1/5
0
0
0
0
0
0